Update unzipEpubFile to check epub mimetype
The epub standard enforces that a file named 'mimetype' exists.  It must
be umcompressed, the first file in the zip archive. The 'mimetype' file
must be a text document in ASCII that contains the string
"application/epub+zip".
This verion of unzipEpubFile check the mimetype, thus ensure the .epub
file is supported by the application.  Note when the file is not
supported, an Exception is thrown, but it is not handled yet. Further
version(s) of unzipEpubFile will handle it, mostly it should notify the
UI of this situation.