Update database book table, add location column
The location column of book table records where the unzipped files and
the backup file of a epub file sit. Also the navigation list needs to
concatenate this location with the navigation_link column of contents
table to form the absolute full path of a resource file.
    This leads to the following changes which are chained together, so I
put them together in one commit.
1. Update database version
2. Update projection map
3. Update create_table SQLite statement