Implement FileIsNotConstructedException class
This exception class is an inner class of EpubFileProcessor class. It
will be thrown when the needed file is not constructed yet, that is a
former method which must be called before this is not called, and the
exception instance should provide enough information to the exception
handler in order to clear the exception.

The public methods now throw this exception instead of any other former
defined exceptions.