Add legacy support on start ProcessEpubFileService
The new version of intent to start ProcessEpubFileService is to set the
Uri from file(usually an epub file) using setData, and that will be all,
the output directory will be a predefined variable. The string extra in
intent now will be ignored, further code should use the new intent
pattern, tests on the service may need an update.