Update NavigationAdapter how the binding happens
The previous version binding is incorrect. Since every time bindView is
called, the View passed in is only created by newView when it is null.
So the proceeding View will reuse the view created before, since the
left padding is get multipled, it will keep getting greater. When it
exceed the width of its parent, it will show as an big empty rectangular
area.
    To retrieve a reliable padding value, do it in newView, since it
will always inflate a fresh new View and the values are original. Then
the values can be used elsewhere without concern.