Stop adding the book if exists when process epub
The simplest approach to this is to stop the service when duplication
found. This is how I did here.
    The method to identify duplication used here is to check if the output
directory is already an existing directory. Thus it only see two epub
file different when they have different file names.
    Future version may go other ways like check the dc:identifier
element textContent in .opf file, which is a more reliable way.