Add navigation order to contents item projection
The Contents.NAVIGATION_ORDER column(retrieved from .ncx file navPoint
playOrder attribute value) is needed to update the mLastOrder per
onNavigationLabelClick(It may be named as 'onNavigationItemClick', but
since it is not, I won't rename it recently). Since the so called last
reading point play order changes as the user navigates.