Use Handler to scroll page when restoring
After I searched and searched, and try various things, I decide I don't
wan't to extend WebView just for a scroll, because there's no explicit
document on Android says there exists a method which you can rely on
called after the WebView is fully loaded and on screen and can be
manipulated. The seemingly solution is to use
WebViewClient.onPageFinished, but don't attempt to do this, this
callback is triggered when the page is fully loaded, but that doesn't
mean it is already on screen, so scrolling here may be swallowed without
notice. Since I only restore the WebView position once every time the
reading activity is started, I will go for a simply solution.
    Instead of dealing with WebView lifecycle methods, I use a Handler
to periodically check WebView.getContentHeight value, if it's greater
than zero, than the WebView can be manipulated and I scroll it, remove
the callback, if not, it will go on and on. This works like a charm!