Add fragment_reading.xml layout file
Note this layout is not inflated by a Fragment, instead, it is inflated
in a ViewStub. The reason I do this is currently all the reading logic
sits under reading activity, only the JavaScript injection is alone. If
I move the WebView and it's control buttons to a Fragment, there will be
too much work, and the UI will looks exactly just the same. So I prefer
not doing so.
There's a problem though, when I put WebView on top of the control
buttons layout, it crashes the application. Anyway, the previous and
next button will work when I finish this serial of commits.