Add ACTION_RESCAN IntentFilter to ProcessEpubFileService
It really sounds weird to let ProcessEpubFileService to handle rescan
action, but for now I just did this. Later I may rename this file, that
is rename the Service. Although this is not recommended, but I'm gonna
do it anyway, since the Service is not exported, and only booklist
activity starts it, I assume it won't be a very big problem.