Implment rescanExistingBooks method
This method clear the database and refill it again by reread necessary
files.
There's one thing worth noticing that the table update order is changed.
This time the book table is updated(insert actually) only when all files
reading have been done.
If you don't do this, the contents table may be still updating when the
user opens the navigation drawer.
This implementation ensures when the book shows in the book list, the
contents table has already finished it's update(insert actually), then
when the user see the book, click on it, it will opens up and the
navigation list in reading activity is ready to go.