Fix ENOTDIR error
This is really weird, because I thought I've already fixed it before.
But it turns out the solution still has flaws. But this time I don't
think there's any problem with the EpubFileProcessor any more.

The problem exists in unzipEpubFile method of EpubFileProcessor.
Sometimes when it starts to unzip the cantainer.xml file, the error
ENOTDIR may occur, I think maybe this has something to do with the
compression algorithm the epub file uses. Because sometimes it seems
that the META-INF directory doesn't exist before META-INF/container.xml
file. So when unzipping the xml file, the error occurs. To solve this,
just make directories leading up the xml file before open it and write
to it.