Implement DeletionConfirmDialog positive button listener
When "Delete" button is clicked, EpubReaderService is started to do the
actual deletion.
Note this will delete rows in database and files on storage.
Note currently the backup file is not deleted, it is reserved for
features like recovery from an accidental deletion. Future versions of
this application may implement that feature.