Implement shouldOverrideUrlLoading of WebViewClient
This method is called every time a link in WebView is clicked, it is
used to determine whether the link should be opened in this
WebView(return false) or by host application(return true), which in the
later situation, ususally handled using an implicit Intent to open it in
another application.
The point to overriding this method is: intercept the url, check if it's
link can be found in contents table, if yes, then update navigation list
current reading position to this item, this allows the previous and next
button to function in a predictable way, that is the buttons will go to
the previous or next page based on the newly open page.