Save current page url and position before jumping
When the user click an in-page link, shouldOverrideUrlLoading is called
before the url will be loaded, save the mBookView.getOriginalUrl here
because it is still the url of the current page. When the user clicks
the go back button, we need the url and position to go to exactly where
the user has left.