Rewrite rescan procedure to show progress bar
Instead of inserting the book as a last step, insert a temporary book
item as soon as the epub processor is constructed, so that the book
item will show as soon as possible. When all files have been read,
update the temporary row will real data.
The progress bar is shown on top of the book item before real data is
inserted into book table during the interval. And this is how I
implement the progress bar when adding a book.