mplement Books.TITLE to ProgressBar binding
The ViewBinder captures the View with id equals
R.id.book_list_item_progress_bar, and compares the title to
TEMPORARY_TITLE_TO_SHOW_PROGRESS_BAR. If it equals, the show the
progress bar, if not, erase it from layout.
This way of binding allows the progress of process a book to be shown
using a progress bar while all added books shown as usual.